fix(datastore): Update for new common constants

+++ datastore/change_metas_v1.go
@@ -63,7 +63,7 @@ func (commonProtocol *CommonProtocol) changeMetasV1(err error, packet nex.Packet
 		}
 
 		// * If the object is pending or rejected, only the owner can interact with it
-		if metaInfo.OwnerID != connection.PID() && (metaInfo.Status == types.UInt8(datastore_constants.DataStatusPending) || metaInfo.Status == types.UInt8(datastore_constants.DataStatusRejected)) {
+		if metaInfo.OwnerID != connection.PID() && (metaInfo.Status == datastore_constants.DataStatusPending || metaInfo.Status == datastore_constants.DataStatusRejected) {
 			pResults = append(pResults, types.NewQResultError(nex.ResultCodes.DataStore.NotFound))
 			continue
 		}
