+++ datastore/get_specific_meta.go
@@ -45,8 +45,8 @@ func (commonProtocol *CommonProtocol) getSpecificMeta(err error, packet nex.Pack
 		}
 
 		// * The owner of an object can always view their objects, but normal users cannot
-		if metaInfo.Status != types.UInt8(datastore_constants.DataStatusNone) && metaInfo.OwnerID != connection.PID() {
-			if metaInfo.Status == types.UInt8(datastore_constants.DataStatusPending) {
+		if metaInfo.Status != datastore_constants.DataStatusNone && metaInfo.OwnerID != connection.PID() {
+			if metaInfo.Status == datastore_constants.DataStatusPending {
 				return nil, nex.NewError(nex.ResultCodes.DataStore.UnderReviewing, "change_error")
 			}
