+++ datastore/prepare_get_object.go
@@ -53,8 +53,8 @@ func (commonProtocol *CommonProtocol) prepareGetObject(err error, packet nex.Pac
 	}
 
 	// * The owner of an object can always view their objects, but normal users cannot
-	if metaInfo.Status != types.UInt8(datastore_constants.DataStatusNone) && metaInfo.OwnerID != connection.PID() {
-		if metaInfo.Status == types.UInt8(datastore_constants.DataStatusPending) {
+	if metaInfo.Status != datastore_constants.DataStatusNone && metaInfo.OwnerID != connection.PID() {
+		if metaInfo.Status == datastore_constants.DataStatusPending {
 			return nil, nex.NewError(nex.ResultCodes.DataStore.UnderReviewing, "change_error")
 		}
 
@@ -63,7 +63,7 @@ func (commonProtocol *CommonProtocol) prepareGetObject(err error, packet nex.Pac
 
 	// TODO - Check param.LockID. See InsertObjectByPreparePostParam for notes on read locks
 
-	notUseFileServer := (metaInfo.Flag & types.UInt32(datastore_constants.DataFlagNotUseFileServer)) != 0
+	notUseFileServer := (metaInfo.Flag & datastore_constants.DataFlagNotUseFileServer) != 0
 	if notUseFileServer {
 		return nil, nex.NewError(nex.ResultCodes.DataStore.InvalidArgument, "PrepareGetObject cannot be used with DataFlagNotUseFileServer")
 	}
